requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
tqdm
pyyaml
pyaml
numpy
scipy
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
polars
loralib==0.0.3
tensorboard==2.13.0
transformers==4.30.2
datasets==2.13.1
pytorch-lightning==2.0.4
deepspeed==0.9.5
rotary-embedding-torch==0.3.5
flash-attn==2.3.4